refactor: api 모듈을 도메인 폴더 구조로 통일
