+++ apps/web/src/api/files/types.ts
@@ -1,0 +1,8 @@
+export interface PresignedUrlResponse {
+  uploadUrl: string;
+  key: string;
+  fileUrl: string;
+  uploadMethod: 'PUT';
+  contentType: string;
+  expiresAt: string;
+}
